Add missing dependency to Num for OCaml >4.06.0
Otherwise the build fails:

    + /usr/bin/ocamlc.opt -c -g -I cil-parser -o nat_big_num.cmo nat_big_num.ml
    File "nat_big_num.ml", line 1, characters 12-19:
    1 | module BI = Big_int
                    ^^^^^^^
    Error: Unbound module Big_int

The recommendation is to switch to Zarith, but for now we can just use
the separate library.

I've tried `-package num` for extract.cmxa as well, but that failed
with:

    + /usr/bin/ocamlopt.opt -a -I /usr/lib64/ocaml/num [...] -o extract.cmxa
    Option -a cannot be used with .cmxa input files.

I didn't figure out where that comes from and what to do about it, so we
add the -I flags manually for now.